refactor : utils 패키지 추가 및 StringFormatter 위치 이동
