+++ app/src/main/java/woowacourse/movie/utils/StringFormatter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.utils
 
 import android.icu.text.DecimalFormat
 import java.time.LocalDate
